streamlit_demo/requirements.txt
# Streamlit Demo Dependencies
streamlit==1.28.0
opencv-python==4.8.1.78
paddleocr==2.7.0.3
paddlepaddle==2.5.0
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
plotly==5.17.0
pillow==10.0.1
python-multipart==0.0.6
pdf2image==1.16.3

# Additional dependencies from main requirements
fastapi==0.104.1
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
aiohttp==3.9.1
transformers>=4.35.0
datasets>=2.14.0
